requirements.txt
python_docx==1.1.2
streamlit==1.41.0
openai==1.57.2
clipboard==0.0.4
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.16.1
st-copy-to-clipboard==0.1.6
cobble==0.1.4
mammoth==1.8.0
pandas==2.2.3
